requirements.txt
django~=4.2.3
djangorestframework